Introduce source ranges and cleanup error handling. (#146)
This introduces a datatype for a range of characters in a source file.
It will be used for embedding source information in the AST for #112 ,
but is currently only being added to info trees.

This PR also removes the extra syntax elements from error handling as
messages contain source information.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.