Support Datetime (#266)
*Issue #, if available:*

*Description of changes:*

This is a pull request that adds abstract definitions of date, datetime
and timedelta for Python.

Datetime is abstractly defined as a pair of (base time, relative
timedelta).
datetime.now() returns (<the curent datetime>, 0).
Adding or subtracting datetime.timedelta updates the relative timedelta
field.
This is co-authored with @andrewmwells-amazon .

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Andrew Wells <anmwells@amazon.com>